requirements.txt
Flask==1.0.2
uWSGI==2.0.17.1
requests==2.18.4
validator.py==1.2.5
